Continuando

+++ src/ar/edu/unq/po2/tpintegrador/AppUsuario.java
@@ -109,7 +109,7 @@ public class AppUsuario implements MovementSensor, Gps {
 		return true;
 	}
 
-	public void aumentarSaldo(int unSaldo) {
+	public void aumentarSaldo(float unSaldo) {
 		saldo = saldo + unSaldo;
 	}
